+++ tests/integration/test_storage_iceberg/test_explanation.py
@@ -1,0 +1,64 @@
+import pytest
+
+from helpers.iceberg_utils import (
+    default_upload_directory,
+    create_iceberg_table,
+    get_uuid_str
+)
+
+from helpers.test_tools import TSV
+
+
+@pytest.mark.parametrize("format_version", ["1", "2"])
+@pytest.mark.parametrize("storage_type", ["s3", "azure", "local"])
+def test_explanation(started_cluster, format_version, storage_type):
+    instance = started_cluster.instances["node1"]
+    spark = started_cluster.spark_session
+    for format in ["Parquet", "ORC", "Avro"]:
+        TABLE_NAME = (
+            "test_explanation_"
+            + format
+            + "_"
+            + format_version
+            + "_"
+            + storage_type
+            + "_"
+            + get_uuid_str()
+        )
+
+        # Types time, timestamptz, fixed are not supported in Spark.
+        spark.sql(
+            f"CREATE TABLE {TABLE_NAME} (x int) USING iceberg TBLPROPERTIES ('format-version' = '{format_version}', 'write.format.default' = '{format}')"
+        )
+
+        spark.sql(f"insert into {TABLE_NAME} select 42")
+        default_upload_directory(
+            started_cluster,
+            storage_type,
+            f"/iceberg_data/default/{TABLE_NAME}/",
+            f"/iceberg_data/default/{TABLE_NAME}/",
+        )
+
+        create_iceberg_table(
+            storage_type, instance, TABLE_NAME, started_cluster, format=format
+        )
+
+        res = instance.query(f"EXPLAIN SELECT * FROM {TABLE_NAME}")
+        res = list(
+            map(
+                lambda x: x.split("\t"),
+                filter(lambda x: len(x) > 0, res.strip().split("\n")),
+            )
+        )
+
+        expected = [
+            [
+                "Expression ((Project names + (Projection + Change column names to column identifiers)))"
+            ],
+            [f"  ReadFromObjectStorage"],
+        ]
+
+        assert res == expected
+
+        # Check that we can parse data
+        instance.query(f"SELECT * FROM {TABLE_NAME}")
