+++ tests/integration/test_storage_iceberg/test_schema_inference.py
@@ -1,0 +1,83 @@
+import pytest
+
+from helpers.iceberg_utils import (
+    default_upload_directory,
+    create_iceberg_table,
+    get_uuid_str
+)
+
+
+
+from helpers.test_tools import TSV
+
+@pytest.mark.parametrize("infer_format", [True, False])
+@pytest.mark.parametrize("format_version", ["1", "2"])
+@pytest.mark.parametrize("storage_type", ["s3", "azure", "local"])
+def test_schema_inference(started_cluster, format_version, storage_type, infer_format):
+    instance = started_cluster.instances["node1"]
+    spark = started_cluster.spark_session
+    for format in ["Parquet", "ORC", "Avro"]:
+        TABLE_NAME = (
+            "test_schema_inference_"
+            + format
+            + "_"
+            + format_version
+            + "_"
+            + storage_type
+            + "_"
+            + get_uuid_str()
+        )
+
+        # Types time, timestamptz, fixed are not supported in Spark.
+        spark.sql(
+            f"CREATE TABLE {TABLE_NAME} (intC int, longC long, floatC float, doubleC double, decimalC1 decimal(10, 3), decimalC2 decimal(20, 10), decimalC3 decimal(38, 30), dateC date,  timestampC timestamp, stringC string, binaryC binary, arrayC1 array<int>, mapC1 map<string, string>, structC1 struct<field1: int, field2: string>, complexC array<struct<field1: map<string, array<map<string, int>>>, field2: struct<field3: int, field4: string>>>) USING iceberg TBLPROPERTIES ('format-version' = '{format_version}', 'write.format.default' = '{format}')"
+        )
+
+        spark.sql(
+            f"insert into {TABLE_NAME} select 42, 4242, 42.42, 4242.4242, decimal(42.42), decimal(42.42), decimal(42.42), date('2020-01-01'), timestamp('2020-01-01 20:00:00'), 'hello', binary('hello'), array(1,2,3), map('key', 'value'), struct(42, 'hello'), array(struct(map('key', array(map('key', 42))), struct(42, 'hello')))"
+        )
+        default_upload_directory(
+            started_cluster,
+            storage_type,
+            f"/iceberg_data/default/{TABLE_NAME}/",
+            f"/iceberg_data/default/{TABLE_NAME}/",
+        )
+
+        if infer_format:
+            create_iceberg_table(
+                storage_type, instance, TABLE_NAME, started_cluster
+            )
+        else:
+            create_iceberg_table(
+                storage_type, instance, TABLE_NAME, started_cluster, format=format
+            )
+        res = instance.query(
+            f"DESC {TABLE_NAME} FORMAT TSVRaw", settings={"print_pretty_type_names": 0}
+        )
+        expected = TSV(
+            [
+                ["intC", "Nullable(Int32)"],
+                ["longC", "Nullable(Int64)"],
+                ["floatC", "Nullable(Float32)"],
+                ["doubleC", "Nullable(Float64)"],
+                ["decimalC1", "Nullable(Decimal(10, 3))"],
+                ["decimalC2", "Nullable(Decimal(20, 10))"],
+                ["decimalC3", "Nullable(Decimal(38, 30))"],
+                ["dateC", "Nullable(Date)"],
+                ["timestampC", "Nullable(DateTime64(6, 'UTC'))"],
+                ["stringC", "Nullable(String)"],
+                ["binaryC", "Nullable(String)"],
+                ["arrayC1", "Array(Nullable(Int32))"],
+                ["mapC1", "Map(String, Nullable(String))"],
+                ["structC1", "Tuple(field1 Nullable(Int32), field2 Nullable(String))"],
+                [
+                    "complexC",
+                    "Array(Tuple(field1 Map(String, Array(Map(String, Nullable(Int32)))), field2 Tuple(field3 Nullable(Int32), field4 Nullable(String))))",
+                ],
+            ]
+        )
+
+        assert res == expected
+
+        # Check that we can parse data
+        instance.query(f"SELECT * FROM {TABLE_NAME}")
